Optimized chunk loading, added limit for max processed chunks per frame

+++ minecraft/chunk_manager.cpp
@@ -9,8 +9,6 @@
 #include "defines.h"
 #include "collision.h"
 
-#include <glm/gtx/norm.hpp>
-
 // Check if target is within Manhattan distance of pos
 bool isWithinSquare(const glm::ivec3& target, const glm::ivec3& pos, const int distance) {
 	return (
@@ -19,11 +17,12 @@ bool isWithinSquare(const glm::ivec3& target, const glm::ivec3& pos, const int d
 		target.z >= pos.z - distance && target.z <= pos.z + distance
 	);
 }
+// Check if target is within euclidean distance of pos
 bool isWithinSphere(const glm::ivec3& target, const glm::ivec3& pos, const int distance) {
-	glm::vec3 rel = (glm::vec3)(target - pos);
-	float len2 = rel.x * rel.x + rel.y * rel.y + rel.z * rel.z;
-	float dist2 = (float)((distance + 1) * (distance + 1));
-	return len2 < dist2 - 0.0001f;
+	glm::ivec3 rel = target - pos;
+	int len2 = rel.x * rel.x + rel.y * rel.y + rel.z * rel.z;
+	int dist2 = (distance + 1) * (distance + 1);
+	return len2 < dist2 ;
 }
 
 ChunkManager::ChunkManager(TerrainGenerator* tg) :
@@ -57,40 +56,76 @@ void ChunkManager::update(const Camera& camera) {
 	// 
 
 	glm::ivec3 chunkPos = blockToChunkPos(camera.getPos()); // Current position in chunk coordinates
+	std::queue<Chunk*> chunkQueue; // Chunks to be sent to the chunk loader queue
 
 	// Reset list
 	renderList.clear();
 
+	// Max number of chunks updated per frame
+	int maxGenerated = 128;
+	int maxLoaded = 128;
+	int maxDeleted = 128;
+
 	// Generate new chunks
-	if (chunkPos != lastChunkPos || camera.renderDistance != lastRenderDistance) {
-		int generateDistance = camera.renderDistance + 1;
+	if (!lastPosLoaded || chunkPos != lastChunkPos || camera.renderDistance != lastRenderDistance) {
+		int distance = camera.renderDistance;
 
-		// Put newly created chunks in shared queue
-		std::unique_lock<std::mutex> mutexLock(chunkLoaderMutex);
+		glm::ivec3 diff = chunkPos - lastChunkPos;
+		glm::ivec3 absDiff = glm::abs(diff);
 
-		lastChunkPos = chunkPos;
-		lastRenderDistance = camera.renderDistance;
-
-		for (int x = chunkPos.x - generateDistance; x <= chunkPos.x + generateDistance; ++x) {
-			for (int y = chunkPos.y - generateDistance; y <= chunkPos.y + generateDistance; ++y) {
-				for (int z = chunkPos.z - generateDistance; z <= chunkPos.z + generateDistance; ++z) {
-					glm::ivec3 pos(x, y, z);
-					Chunk* c = getChunk(pos);
-					if (c == nullptr) {
-						c = createChunk(pos);
-						chunkLoaderQueue.push(c);
-					} else {
-						if (c->isLoaded && !c->meshGenerated && isWithinSphere(c->pos, chunkPos, camera.renderDistance)) {
-							c->isLoaded = false;
-							chunkLoaderQueue.push(c);
+		if (!lastPosLoaded || camera.renderDistance != lastRenderDistance || absDiff.x + absDiff.y + absDiff.z >= 2 * distance + 1) {
+			// Loop through all chunks in range
+			glm::ivec3 pos;
+			for (pos.x = chunkPos.x - distance; pos.x <= chunkPos.x + distance; ++pos.x) {
+				for (pos.y = chunkPos.y - distance; pos.y <= chunkPos.y + distance; ++pos.y) {
+					for (pos.z = chunkPos.z - distance; pos.z <= chunkPos.z + distance; ++pos.z) {
+						if (getChunk(pos) == nullptr) {
+							if (maxGenerated > 0) {
+								--maxGenerated;
+								chunkQueue.push(createChunk(pos));
+							} else {
+								lastPosLoaded = false;
+								goto break_loop;
+							}
 						}
 					}
 				}
 			}
-		}
+			lastChunkPos = chunkPos;
+			lastPosLoaded = true;
+		} else {
+			// Loop only through chunks that weren't in range of last loaded position
+			for (int i = 0; i < 3; ++i) {
+				if (diff[i] != 0) {
+					glm::ivec3 min = chunkPos - distance;
+					glm::ivec3 max = chunkPos + distance;
+					if (diff[i] > 0) min[i] = max[i] - diff[i] + 1;
+					else if (diff[i] < 0) max[i] = min[i] - diff[i] - 1;
 
-		mutexLock.unlock();
-		chunkLoaderCondition.notify_all();
+					glm::ivec3 pos;
+					for (pos.x = min.x ; pos.x <= max.x; ++pos.x) {
+						for (pos.y = min.y; pos.y <= max.y; ++pos.y) {
+							for (pos.z = min.z; pos.z <= max.z; ++pos.z) {
+								if (getChunk(pos) == nullptr) {
+									if (maxGenerated > 0) {
+										--maxGenerated;
+										chunkQueue.push(createChunk(pos));
+									} else {
+										lastPosLoaded = false;
+										goto break_loop;
+									}
+								}
+							}
+						}
+					}
+				}
+			}
+			lastChunkPos = chunkPos;
+			lastPosLoaded = true;
+		}
+		// Break outer loop
+	break_loop:
+		lastRenderDistance = camera.renderDistance;
 	}
 
 	// Process existing chunks
@@ -99,35 +134,75 @@ void ChunkManager::update(const Camera& camera) {
 		Chunk* c = it->second;
 		if (c->isLoaded) { // Make sure chunk isn't being processed by secondary thread
 			if (isWithinSphere(c->pos, chunkPos, camera.renderDistance)) { // Chunk in render distance
+				// Generate mesh
+				if (!c->meshGenerated) {
+					if (maxGenerated > 0) {
+						--maxGenerated;
+						c->isLoaded = false;
+						chunkQueue.push(c);
+					}
+				}
 				// Load mesh
-				if (c->meshGenerated && !c->meshLoaded) {
-					c->loadMesh();
+				else if (c->meshGenerated && !c->meshLoaded) {
+					if (maxLoaded > 0) {
+						--maxLoaded;
+						c->loadMesh();
+					}
 				}
 				// Add to render queue
-				if (c->meshLoaded) {
+				if (c->meshLoaded && (c->n_indices[0] > 0 || c->n_indices[1] > 0)) {
 					AABB aabb(c);
 					if (frustum.intersects(aabb)) {
 						renderList.push_back(c);
 					}
 				}
 			} else if (!isWithinSquare(c->pos, chunkPos, camera.renderDistance + 2)) { // Chunk outside render distance + 2
-				// Delete chunk if possible
-				bool safeToDelete = true;
-				for (int side = 0; side < 6; ++side) {
-					if (c->neighbors[side] != nullptr && !c->neighbors[side]->isLoaded) {
-						safeToDelete = false;
-						break;
+				if (maxDeleted > 0) {
+					--maxDeleted;
+
+					// Delete chunk if possible
+					bool safeToDelete = true;
+					for (int side = 0; side < 6; ++side) {
+						if (c->neighbors[side] != nullptr && !c->neighbors[side]->isLoaded) {
+							safeToDelete = false;
+							break;
+						}
 					}
-				}
-				if (safeToDelete) {
-					it = deleteChunk(it);
-					// Don't increment iterator
-					continue;
+					if (safeToDelete) {
+						it = deleteChunk(it);
+						// Don't increment iterator
+						continue;
+					}
 				}
 			}
 		}
 		// Increment iterator
 		++it;
+	}
+
+	// Send chunks to the chunk loader queue
+	if (!chunkQueue.empty()) {
+		std::unique_lock<std::mutex> mutexLock(chunkLoaderMutex); // Lock mutex
+
+		while (!chunkQueue.empty()) {
+			// Move chunks from one queue to another
+			Chunk* c = chunkQueue.front();
+			chunkQueue.pop();
+			chunkLoaderQueue.push(c);
+			// If a chunk has a missing neighbor, create it and add it to the queue as well
+			for (int side = 0; side < 6; ++side) {
+				if (c->neighbors[side] == nullptr) {
+					c->neighbors[side] = createChunk(getNeighborPos(side, c->pos));
+					chunkLoaderQueue.push(c->neighbors[side]);
+				}
+			}
+		}
+		// Send position and render distance to chunk loader
+		chunkLoaderPos = chunkPos;
+		chunkLoaderDistance = camera.renderDistance;
+
+		mutexLock.unlock(); // Unlock mutex
+		chunkLoaderCondition.notify_all();
 	}
 }
 
@@ -184,30 +259,28 @@ void ChunkManager::setBlock(const glm::ivec3& blockPos, int blockType) {
 	}
 }
 
+// Create new chunk, link existing neighbors and insert it into chunk map
 Chunk* ChunkManager::createChunk(const glm::ivec3& chunkPos) {
 	Chunk* c = new Chunk(chunkPos);
-
+	// Link existing neighbors
 	for (int side = 0; side < 6; ++side) {
-		glm::ivec3 chunkPos2 = chunkPos + glm::ivec3(
-			neighborsIndices[3 * side],
-			neighborsIndices[3 * side + 1],
-			neighborsIndices[3 * side + 2]
-		);
+		glm::ivec3 neighborPos = getNeighborPos(side, chunkPos);
 
-		Chunk* c2 = getChunk(chunkPos2);
-		if (c2 != nullptr) {
-			int side2 = (side % 2 == 0) ? side + 1 : side - 1;
-			c->neighbors[side] = c2;
-			c2->neighbors[side2] = c;
+		Chunk* cn = getChunk(neighborPos);
+		if (cn != nullptr) {
+			int neighborSide = (side % 2 == 0) ? side + 1 : side - 1;
+			c->neighbors[side] = cn;
+			cn->neighbors[neighborSide] = c;
 		}
 	}
-
+	// Put chunk in map
 	chunks.insert(std::make_pair(chunkPos, c));
 	//chunks[glm::ivec3(x, y, z)] = c;
+
 	return c;
 }
 
-ChunkMap::iterator ChunkManager::deleteChunk(ChunkMap::iterator it) {
+ChunkMap::iterator ChunkManager::deleteChunk(const ChunkMap::iterator& it) {
 	Chunk* c = it->second;
 	delete c;
 	return chunks.erase(it);
@@ -227,26 +300,28 @@ void chunkLoaderFunc(ChunkManager* cm) {
 		if (!cm->chunkLoaderIsRunning)
 			break;
 
-		glm::ivec3 chunkPos = cm->lastChunkPos;
+		glm::ivec3 chunkPos = cm->chunkLoaderPos;
+		int renderDistance = cm->chunkLoaderDistance;
 		TerrainGenerator* terrainGen = cm->terrainGen;
 
+		// Get new chunks
+		do {
+			/*Chunk* c = cm->chunkLoaderQueue.front();
+			cm->chunkLoaderQueue.pop();
+			dataQueue.push(c);*/
+			dataQueue.swap(cm->chunkLoaderQueue);
+		} while (!cm->chunkLoaderQueue.empty());
+
+		mutexLock.unlock();
+
 		// Priority queue that loads meshes in order of increasing distance from player
-		auto cmp = [&chunkPos](Chunk* &left, Chunk* &right) {\
+		auto cmp = [&chunkPos](Chunk*& left, Chunk*& right) {
 			glm::vec3 l = left->pos - chunkPos;
 			glm::vec3 r = right->pos - chunkPos;
 			//return (l.x * l.x + l.y * l.y + l.z * l.z) > (r.x * r.x + r.y * r.y + r.z * r.z); // Euclidean distance
 			return (abs(l.x) + abs(l.y) + abs(l.z)) > (abs(r.x) + abs(r.y) + abs(r.z)); // Manhattan distance
 		};
 		std::priority_queue<Chunk*, std::vector<Chunk*>, decltype(cmp)> meshQueue(cmp);
-
-		// Get new chunks
-		do {
-			Chunk* c = cm->chunkLoaderQueue.front();
-			cm->chunkLoaderQueue.pop();
-			dataQueue.push(c);
-		} while (!cm->chunkLoaderQueue.empty());
-
-		mutexLock.unlock();
 
 		// Generate data for new chunks
 		while (!dataQueue.empty()) {
@@ -256,7 +331,7 @@ void chunkLoaderFunc(ChunkManager* cm) {
 			if (!c->dataGenerated)
 				c->generateData(terrainGen);
 
-			if (isWithinSphere(c->pos, chunkPos, cm->lastRenderDistance))
+			if (c->hasAllNeighbors() && isWithinSphere(c->pos, chunkPos, renderDistance))
 				meshQueue.push(c);
 			else
 				c->isLoaded = true;
